Poly-commitment: move combine_evaluations in kzg module
The routine isn't used for IPA, it is only used by the KZG PCS.
